dev: remove unwraps from crates/pallet and crates/primitives
