+++ crates/pallets/starknet/src/blockifier_state_adapter.rs
@@ -164,7 +164,7 @@ impl<T: Config> State for BlockifierStateAdapter<T> {
 
     fn set_contract_class(&mut self, class_hash: &ClassHash, contract_class: ContractClass) -> StateResult<()> {
         let class_hash: ClassHashWrapper = class_hash.0.into();
-        let contract_class: ContractClassWrapper = ContractClassWrapper::try_from(contract_class).unwrap();
+        let contract_class: ContractClassWrapper = ContractClassWrapper::from(contract_class);
 
         crate::ContractClasses::<T>::insert(class_hash, contract_class);
